Replace ReportVersions::Tiny with Test::ReportPrereqs
`ReportVersions::Tiny` is deprecated; `Test::ReportPrereqs` is the
recommended alternative.  This change replaces the deprecated plugin with
the recommended plugin in `dist.ini`.  Although this replacement had been
made in version 0.0.5, the change in `dist.ini` had been missed which was
causing users of the bundle to still use `ReportVersions::Tiny`.  It could,
however, be possible that the issue is actually caused by the 0.0.4 version
being considered by CPAN to be the version to use rathern than the 0.0.5
version.  Nevertheless, this change, coupled with a new release, would help
many downstream users as well as close RT issue #103820.